Add mousedown event
I also had to switch to straight up asynchronous testing since we have to
wait for the images to load up before initializing the knob.

There's one failing test, but I can't figure out why. Logging the value
shows that it's correct, but the expectations are wrong. Ignoring it for
now since it doesn't seem to have any actual consequences.